Use single staticlib
