+++ src/rust/Cargo.toml
@@ -3,6 +3,7 @@ members = [
     "crates/ain-evm",
     "crates/ain-evm-state",
     "crates/ain-evm-runtime",
+    "crates/ain-evm-ffi",
     "crates/ain-grpc",
 ]
