+++ src/rust/Makefile
@@ -25,24 +25,19 @@ endif
 all: build-evm-pkg build-grpc-pkg
 
 build-evm-pkg:
-	CRATE_CC_NO_DEFAULTS=1 $(CARGO) build --package ain-evm --release $(if $(TARGET),--target $(TARGET),)
+	CRATE_CC_NO_DEFAULTS=1 $(CARGO) build --package ain-evm-ffi --release $(if $(TARGET),--target $(TARGET),)
 	mkdir -p include
-	cp target/$(if $(TARGET),$(TARGET)/,)release/libain_evm.a  libain_evm.a
+	cp target/$(if $(TARGET),$(TARGET)/,)release/libain_evm_ffi.a libain_evm_ffi.a
 	cp target/libain_evm.h include/
 	cp target/libain_evm.cpp libain_evm.cpp
 
 build-grpc-pkg:
 	CRATE_CC_NO_DEFAULTS=1 $(CARGO) build --package ain-grpc --release $(if $(TARGET),--target $(TARGET),)
-	mkdir -p include
-	cp target/$(if $(TARGET),$(TARGET)/,)release/libain_grpc.a  libain_grpc.a
-	cp target/libain_grpc.h include/
-	cp target/libain_grpc.cpp libain_grpc.cpp
 
 clean:
 	$(CARGO) clean
 	rm -rf include
-	rm -rf libain_evm.a libain_evm.cpp
-	rm -rf libain_grpc.a libain_grpc.cpp
+	rm -rf libain_evm_ffi.a libain_evm.cpp
 
 distclean: clean
