+++ src/rust/crates/ain-evm-ffi/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "ain-evm-ffi"
+version = "0.1.0"
+edition = "2021"
+build = "build.rs"
+
+[lib]
+crate-type = ["staticlib"]
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+ain-evm-state = { path = "../ain-evm-state" }
+ain-evm-runtime = { path = "../ain-evm-runtime" }
+ain-evm = { path = "../ain-evm" }
+ain-grpc = { path = "../ain-grpc" }
+
+# Build
+cxx = "1.0"
+
+[build-dependencies]
+substrate-build-script-utils = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.27" }
+cxx-gen = "0.7"
+proc-macro2 = "1.0"
