+++ src/rust/crates/ain-evm/Cargo.toml
@@ -2,23 +2,13 @@
 name = "ain-evm"
 version = "0.1.0"
 edition = "2021"
-build = "build.rs"
-
-[lib]
-crate-type = ["staticlib"]
 
 [dependencies]
 evm = { version = "0.37", default-features = false }
 primitive-types = "0.12.1"
 log = "0.4"
 
-# Build
-cxx = "1.0"
-
 ain-evm-runtime = { path = "../ain-evm-runtime" }
 ain-evm-state = { path = "../ain-evm-state" }
+ain-grpc = { path = "../ain-grpc" }
 
-[build-dependencies]
-substrate-build-script-utils = { git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.27" }
-cxx-gen = "0.7"
-proc-macro2 = "1.0"
